+++ src/server/services/lifetimeStats/index.ts
@@ -1,9 +1,6 @@
 import { SERVICE_Call__getAllReposWithCommitCounts } from '@/server/services/reposAndCommitCounts'
-import {
-    SHARED_APIFields__LifetimeStats,
-    SHARED_Model__AllLanguageStats,
-    SHARED_Model__LifetimeStats,
-} from '@/shared/models'
+import { SHARED_APIFields__LifetimeStats, SHARED_Model__LifetimeStats } from '@/shared/models'
+import { SHARED_Model__AllLanguageStats } from '@/shared/models/models/Language'
 import {
     SHARED_Model__RepoCommitCountStats,
     SHARED_Model__RepoWithCommitCountsAndLanguages,
