+++ src/shared/models/models/Language.ts
@@ -1,0 +1,10 @@
+export interface SHARED_Model__Language {
+    size: number
+    color: string
+    name: string
+}
+
+export interface SHARED_Model__AllLanguageStats {
+    totalDiskUsage: number
+    allLanguageData: SHARED_Model__Language[]
+}
